Subtitle Refactoring :
- Add a text directory
- Add a proxy class for text buffer (it will be used for each type of text)
- Extract HTML render for embedded text, from TextSourceBuffer
