+++ src/streaming/MediaPlayer.js
@@ -36,10 +36,10 @@ import ManifestLoader from './ManifestLoader';
 import LiveEdgeFinder from './utils/LiveEdgeFinder';
 import ErrorHandler from './utils/ErrorHandler';
 import Capabilities from './utils/Capabilities';
-import TextTracks from './TextTracks';
+import TextTracks from './text/TextTracks';
 import SourceBufferController from './controllers/SourceBufferController';
 import RequestModifier from './utils/RequestModifier';
-import TextSourceBuffer from './TextSourceBuffer';
+import TextSourceBuffer from './text/TextSourceBuffer';
 import URIQueryAndFragmentModel from './models/URIQueryAndFragmentModel';
 import ManifestModel from './models/ManifestModel';
 import MediaPlayerModel from './models/MediaPlayerModel';
